pyproject.toml
[tool.black]
line-length = 120
target-version = ['py310']
force-exclude = '''
(
    \.eggs
  | \.git
  | build
  | dist
  | venv
  | .venv
  | migrations
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
skip_glob = ['**/migrations/**', '**/.venv/**']


[tool.poetry]
name = "empha-api"
version = "0.1.0"
description = ""
authors = ["fayzikuloff.b <fayzikuloff.b@gmail.com>"]
readme = "README.md"
packages = [{include = "empha_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
python-dotenv = "^0.21.1"
psycopg2-binary = "^2.9.5"
drf-spectacular = {extras = ["sidecar"], version = "^0.25.1"}
django-dirtyfields = "^1.9.1"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.2.2"}


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
